rename hooks file to useFuse for better searching
